+++ app/Modules/Accounting/Controllers/RefundController.php
@@ -123,7 +123,7 @@ class RefundController extends Controller
     public function set_product(RefundProduct $product, Request $request): RedirectResponse
     {
         try {
-            $this->service->setProduct($product, $request->integer('quantity'));
+            $this->service->setProduct($product, $request->float('quantity'));
             return redirect()->back()->with('success', 'Сохранено');
         } catch (\DomainException $e) {
             return redirect()->back()->with('error', $e->getMessage());
